refact: cli clean up
