+++ cli/src/lib.rs
@@ -1,6 +1,3 @@
 pub mod commands;
-pub mod common;
 pub mod interactive;
 pub mod query;
-
-pub use common::run;
